refactor(app): extract MetricsView into modular hero/charts/stats tree
Phase 0 (B5) of the mode-adaptive dashboard — pure refactor, no behaviour
change (147 app tests pass, type-check clean). Splits the 1285-line
MetricsView into:
- hero/: RateFidelityCard, ThroughputTwinCard, ErrorRateCard,
  DroppedRequestsCard + HeroRow router (mode→2 cards + universal Error Rate).
- charts/: ThroughputOverTimeChart, HdrPercentilePlot, TimingWaterfall +
  the relocated Latency/Percentiles charts. All consume shared
  TIME_SERIES_DIMS/HDR_DIMS from chartGeometry (gate #4).
- stats/: StatCard + ModeStatsRow router (mode→4 stats).
- MetricsView becomes a thin orchestrator: useMode + one memoized
  DashboardDerived bundle, passed to HeroRow/ModeStatsRow.

HeroRow/ModeStatsRow render the constant_rps layout for every mode until
the fan-out tasks (A1/A2) add the per-mode branches — so the experts each
own their own file with zero MetricsView contention. Plumbs configured
`concurrency` + progress fields through for A1/A2.
